@@ -3,14 +3,13 @@ name = flagship
 version = attr: flagship.__version__
 description = Flagship Python SDK
 long_description = file: README.md
-long-description-content-type = text/markdown
+long_description_content_type = text/markdown
 license = Apache License Version 2.0
 license_files = file: LICENSE
 classifiers =
     License :: OSI Approved :: Apache Software License
     Topic :: Software Development :: Libraries :: Python Modules
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -23,17 +22,17 @@ project_urls =
     Sources = https://github.com/abtasty/flagship-python-sdk
     Documentation = https://docs.developers.flagship.io/docs/python-v2-1
 [options]
-python_requires = >=2.7
+python_requires = >=3.7
 packages = find:
 package_dir =
     =.
 
 install_requires =
-    requests ==2.27.1
-    typing ==3.7.4.3 ; python_version < '3.5'
-    six ==1.16.0
+    requests >=2.31.0
+    pytest >=4.6.11
     enum34 ~=1.1.10
-    responses ==0.17.0
+    responses >=0.17.0
+    mock>=3.0.5
 
 [options.packages.find]
 where = .
