@@ -1,48 +1,30 @@
-# coding: utf8
-from datetime import date, datetime
-import random
-import json
+import asyncio
 import sys
+import time
 
-from flagship.app import Flagship
-from flagship.cache.cache_visitor import VisitorCacheManager
-from flagship.config import Config
-from flagship.handler import FlagshipEventHandler
-from flagship.helpers.hits import Page, Screen
+from flagship import *
+from flagship.cache_manager import SqliteCacheManager
+from flagship.config import DecisionApi, Bucketing
+from flagship.hits import Screen
+from flagship.tracking_manager import TrackingManagerConfig
 
-
-class CustomEventHandler(FlagshipEventHandler):
-    def __init__(self):
-        FlagshipEventHandler.__init__(self)
-
-    def on_log(self, level, message):
-        FlagshipEventHandler.on_log(self, level, ">>> " + message)
-        pass
-
-    def on_exception_raised(self, exception, traceback):
-        FlagshipEventHandler.on_exception_raised(self, exception, traceback)
-        pass
-
-
-class CustomVisitorCacheManager(VisitorCacheManager):
-
-    def save(self, visitor_id, visitor_data):
-        print("VisitorCacheManager look up " + json.dumps(visitor_data))
-
-    def lookup(self, visitor_id):
-        print("VisitorCacheManager look up " + visitor_id)
-        return None
 
 def init():
     print(sys.version)
-    t = CustomEventHandler()
+    Flagship.start('__env_id__', '__api_key__', DecisionApi(timeout=3000,
+                                                            cache_manager=SqliteCacheManager(),
+                                                            log_level=LogLevel.ALL,
+                                                            tracking_manager_config=TrackingManagerConfig(
+                                                                time_interval=10000,
+                                                                max_pool_size=5)))  ## Demo
 
-    Flagship.instance().start("_my_env_id", "_my_api_key_",
-                              Config(event_handler=t, mode=Config.Mode.BUCKETING, polling_interval=5, timeout=0.1,
-                                     visitor_cache_manager=CustomVisitorCacheManager()))
-    v = Flagship.instance().create_visitor("visitorId_1", True, {'isVIPUser': True, 'daysSinceLastLaunch': 3})
-    v.synchronize_modifications()
-    value = v.get_modification("target", "default", True)
-    v.send_hit(Screen("python screen view"))
-    print(value)
+    visitor = Flagship.new_visitor('visitor-A', context={'testing_tracking_manager': True})
+    visitor.fetch_flags()
+    visitor.get_flag("my_flag", 'default').value()
+    visitor.send_hit(Screen("screen 1"))
+    time.sleep(2)
+
+
+
+
 init()
